backend/pyproject.toml
[project]
name = "strata-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Fledge Shiu", email = "xzk0701@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "alembic==1.16.5",
    "asyncpg==0.30.0",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastapi[all]==0.116.2",
    "greenlet==3.2.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv==1.1.1",
    "returns==0.26.0",
    "sqlalchemy[asyncio]==2.0.43",
]

[tool.uv.sources]
strata-ai = { workspace = true }

[dependency-groups]
dev = [
    "aiosqlite==0.21.0",
    "pyright==1.1.405",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "strata-ai",
]

[project.scripts]
strata-ai = "strata_ai:main"

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[tool.pyright]
venvPath = '.'
venv = ".venv"
exclude = ['.venv']

[tool.pytest.ini_options]
asyncio_mode = "auto"
